Assignments/A3-KafkaTask/jobs/requirements.txt
kafka-python==2.1.5
py4j==0.10.9.7
pyspark==3.5.5
notebook==7.4.2